add analysis test for installer dependencies
We want to avoid having a depdendency on the helper built from source (and a Go toolchain) if a prebuilt can be used.
This analysis test uses an aspect to walk the depdendency graph in order to check if the source built helper is in the transitive closure of the installer.