github: fix auth header for githubusercontent
weirdly, the scheme for accessing single artifacts in a repository tree
via raw.githubusercontent.com is slightly different than the method for
github.com itself. github being github, i can't find this written down
in any official documentation, but you can see [questions][1], [discussions][2],
and [gists][3] all over the place explaining it. there are two ways of
doing it, apparently, but the one i discovered as working first is to
use your github token as the user component of the Basic auth scheme.
looks like you may also be able to use the bare token by specifying the
nonstandard Token scheme in the auth header, but it seems like six of
one, half dozen of the other.

[1]: https://github.com/orgs/community/discussions/160828
[2]: https://github.com/orgs/community/discussions/22537
[3]: https://gist.github.com/jcefoli/4f24620acc0521f895649c933ee6685b