github: check config file after any lookup failure
currently, if the default keyring lookup fails for any reason other
than not found (like the keyring couldn't be reached), the github
provider won't check the config file. i don't see why it shouldn't- the
current behavior means that the config file doesn't get checked at all
on linux machines that aren't running a desktop environment or don't
have dbus-launch in the path. we'd like that to happen, while also
checking the keyring on mac.